refactor: implement file config store
