+++ ts/packages/node/src/stores/JsonFile.ts
@@ -1,44 +1,29 @@
-import { promises as fs } from 'fs';
-import _ from 'lodash';
-import { ConfigStore, type ConfigStoreQuery, type Config, Convert } from '@configu/ts';
+import { promises as fs, existsSync } from 'fs';
+import { type Config, Convert } from '@configu/ts';
+import { FileConfigStore } from './FileConfigStore';
 
 export type JsonFileConfigStoreConfiguration = { path: string };
 
-export class JsonFileConfigStore extends ConfigStore {
-  private readonly path: string;
+export class JsonFileConfigStore extends FileConfigStore {
   constructor({ path }: JsonFileConfigStoreConfiguration) {
-    super('json-file');
-    this.path = path;
+    super('json-file', path);
   }
 
-  async read(): Promise<Config[]> {
+  async init() {
+    const fileExists = await existsSync(this.path);
+    if (!fileExists) {
+      const initialFileState = Convert.configStoreContentsToJson([]);
+      await fs.writeFile(this.path, initialFileState);
+    }
+  }
+
+  async read() {
     const data = await fs.readFile(this.path, 'utf8');
     return Convert.toConfigStoreContents(data);
   }
 
-  async write(nextConfigs: Config[]): Promise<void> {
+  async write(nextConfigs: Config[]) {
     const data = Convert.configStoreContentsToJson(nextConfigs);
     await fs.writeFile(this.path, data);
-  }
-
-  async get(queries: ConfigStoreQuery[]): Promise<Config[]> {
-    const storedConfigs = await this.read();
-
-    return storedConfigs.filter((config) => {
-      return queries.some(({ set, key }) => {
-        return set === config.set && key === config.key;
-      });
-    });
-  }
-
-  async set(configs: Config[]): Promise<void> {
-    const storedConfigs = await this.read();
-
-    const nextConfigs = _([...configs, ...storedConfigs])
-      .uniqBy((config) => `${config.set}.${config.key}`)
-      .filter((config) => Boolean(config.value))
-      .value();
-
-    await this.write(nextConfigs);
   }
 }
